pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ksef-client-python"
version = "0.1.0"
description = "KSeF API client for Python"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "KSeF Python SDK Contributors"}]
keywords = ["ksef", "e-invoice", "poland", "api"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "httpx>=0.27.0",
  "cryptography>=42.0.0",
]

[project.optional-dependencies]
xml = [
  "lxml>=5.0.0",
  "xmlsec>=1.3.13",
]
qr = [
  "qrcode[pil]>=7.4.2",
  "pillow>=10.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ksef_client = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["ksef_client"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
